requirements.txt
pycryptodome
pyasn1_modules


